+++ adr/__init__.py
@@ -1,3 +1,3 @@
 """Automatic Disc Ripper: Automatic Disc Ripper."""
 
-__version__ = "1.28.0"
+__version__ = "1.29.0"